fix: use fs sync method to write files
In certain cases the stats are not written, this seems to be because the process is being terminated before it finish to write. The reason being is using a sync version of tap.

To get around this, the sync method of fs should be used to allow the creation of files.

Related to: https://github.com/angular/angular-cli/issues/12763